Add an application controller for internal controllers
