+++ railties/lib/rails/application_controller.rb
@@ -1,0 +1,16 @@
+class Rails::ApplicationController < ActionController::Base # :nodoc:
+  self.view_paths = File.expand_path('../templates', __FILE__)
+  layout 'application'
+
+  protected
+
+  def require_local!
+    unless local_request?
+      render text: '<p>For security purposes, this information is only available to local requests.</p>', status: :forbidden
+    end
+  end
+
+  def local_request?
+    Rails.application.config.consider_all_requests_local || request.local?
+  end
+end
