+++ railties/lib/rails/info_controller.rb
@@ -1,9 +1,9 @@
+require 'rails/application_controller'
 require 'action_dispatch/routing/inspector'
 
-class Rails::InfoController < ActionController::Base # :nodoc:
-  self.view_paths = File.expand_path('../templates', __FILE__)
+class Rails::InfoController < Rails::ApplicationController # :nodoc:
   prepend_view_path ActionDispatch::DebugExceptions::RESCUES_TEMPLATE_PATH
-  layout -> { request.xhr? ? nil : 'application' }
+  layout -> { request.xhr? ? false : 'application' }
 
   before_filter :require_local!
 
@@ -19,17 +19,5 @@ class Rails::InfoController < ActionController::Base # :nodoc:
   def routes
     @routes_inspector = ActionDispatch::Routing::RoutesInspector.new(_routes.routes)
     @page_title = 'Routes'
-  end
-
-  protected
-
-  def require_local!
-    unless local_request?
-      render text: '<p>For security purposes, this information is only available to local requests.</p>', status: :forbidden
-    end
-  end
-
-  def local_request?
-    Rails.application.config.consider_all_requests_local || request.local?
   end
 end
